requirements.txt
qiskit_nature==0.1.1
qiskit_terra<0.18.0
qiskit_ignis==0.6.0
qiskit-aer==0.8.1
qiskit_ibmq_provider==0.12.3
pyscf==1.7.6
h5py==3.1
matplotlib>=2.1,<3.4
jupyter
